+++ test/3-deferred-configs.js
@@ -49,6 +49,16 @@ vows.describe('Tests for deferred values - JavaScript').addBatch({
     "second defer function return original value." : function () {
       assert.equal(CONFIG.original.deferredOriginal, undefined);
     },
+
+    "defer functions resolved within an array" : function () {
+      assert.equal(CONFIG.list[2], 3);
+      assert.equal(CONFIG.fromList, 6);
+    },
+
+    "defer functions execution order resolves when accessing other defer functions" : function () {
+      assert.equal(CONFIG.a, 'my this is 6!');
+      assert.equal(CONFIG.c, 'my this is 6! this is 6!');
+    },
   }
 })
 .export(module);