+++ test/x-config-ts-module-exports/local.ts
@@ -1,0 +1,7 @@
+import { Config } from './types';
+
+const localConfig: Partial<Config> = {
+ siteTitle : 'New Instance!',
+};
+
+module.exports = localConfig;