+++ test/x-config-ts-module-exports/types.ts
@@ -1,0 +1,5 @@
+export interface Config {
+  siteTitle: string;
+  latitude: number;
+  longitude: number;
+};